+++ app/dashboard/transactions/client.tsx
@@ -1,0 +1,140 @@
+"use client";
+import { DataTable } from "@/components/data-table";
+import { Button } from "@/components/ui/button";
+import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
+import { Skeleton } from "@/components/ui/skeleton";
+import { transactions as transactionSchema } from "@/db/schema";
+import { useSelectAccount } from "@/features/accounts/hooks/use-select-account";
+import { useBulkCreateTransactions } from "@/features/transactions/api/use-bulk-create-transactions";
+import { useBulkDeleteTransactions } from "@/features/transactions/api/use-bulk-delete-transactions";
+import { useGetTransactions } from "@/features/transactions/api/use-get-transtractions";
+import { useNewTransaction } from "@/features/transactions/hooks/use-new-transaction";
+import { Loader2Icon, PlusIcon } from "lucide-react";
+import { useTranslations } from "next-intl";
+import { useState } from "react";
+import { toast } from "sonner";
+import { columns } from "./columns";
+import ImportCard from "./import-card";
+import UploadButton from "./upload-button";
+
+enum VARIANTS {
+  LIST = "LIST",
+  IMPORT = "IMPORT",
+}
+
+const INITIAL_IMPORT_RESULTS = {
+  data: [],
+  error: null,
+  meta: {},
+};
+
+const TransactionsClient = () => {
+  const t = useTranslations();
+
+  const [AccountDialog, confirm] = useSelectAccount();
+  const [variant, setVariant] = useState<VARIANTS>(VARIANTS.LIST);
+  const [importResults, setImportResults] = useState(INITIAL_IMPORT_RESULTS);
+
+  const onUpload = (result: typeof INITIAL_IMPORT_RESULTS) => {
+    setImportResults(result);
+    setVariant(VARIANTS.IMPORT);
+  };
+
+  const onCancelImport = () => {
+    setImportResults(INITIAL_IMPORT_RESULTS);
+    setVariant(VARIANTS.LIST);
+  };
+
+  const newTransaction = useNewTransaction();
+  const createTransactions = useBulkCreateTransactions();
+  const transactionsQuery = useGetTransactions();
+  const deleteTransactions = useBulkDeleteTransactions();
+  const transactions = transactionsQuery.data || [];
+
+  const isDisabled =
+    transactionsQuery.isLoading || deleteTransactions.isPending;
+
+  if (transactionsQuery.isLoading) {
+    return (
+      <Card className="border-none drop-shadow-sm">
+        <CardHeader>
+          <Skeleton className="h-8 w-48" />
+        </CardHeader>
+        <CardContent>
+          <div className="h-[500px] w-full flex items-center justify-center">
+            <Loader2Icon className="size-6 text-slate-300 animate-spin" />
+          </div>
+        </CardContent>
+      </Card>
+    );
+  }
+
+  const onSubmitImport = async (
+    values: (typeof transactionSchema.$inferInsert)[]
+  ) => {
+    const accountId = await confirm();
+    if (!accountId) {
+      return toast.error(t("Toast.ImportFailure"));
+    }
+    const data = values.map((value) => ({
+      ...value,
+      accountId: accountId as string,
+    }));
+
+    createTransactions.mutate(data, {
+      onSuccess: () => {
+        toast.success(t("Toast.Success"));
+        onCancelImport();
+      },
+    });
+  };
+
+  if (variant === VARIANTS.IMPORT) {
+    return (
+      <>
+        <AccountDialog />
+        <ImportCard
+          data={importResults.data}
+          onCancel={onCancelImport}
+          onSubmit={onSubmitImport}
+        />
+      </>
+    );
+  }
+
+  return (
+    <Card className="border-none drop-shadow-sm">
+      <CardHeader className="gap-y-2 lg:flex-row lg:items-center lg:justify-between">
+        <CardTitle className="text-xl line-clamp-1">
+          {t("Common.Page.Header", { key: t("TransactionsPage.Header") })}
+        </CardTitle>
+        <div className="flex flex-col lg:flex-row gap-y-2 items-center gap-x-2">
+          <Button
+            size={"sm"}
+            onClick={newTransaction.onOpen}
+            className="w-full lg:w-auto"
+          >
+            <PlusIcon className="size-4 mr-2" />
+            {t("Common.Action.New")}
+          </Button>
+          <UploadButton onUpload={onUpload} />
+        </div>
+      </CardHeader>
+      <CardContent>
+        <DataTable
+          columns={columns(t)}
+          data={transactions}
+          filterKey="payee"
+          filterKeyTranslate={t("TransactionsPage.Column.Payee")}
+          onDelete={(row) => {
+            const ids = row.map((r) => r.original.id);
+            deleteTransactions.mutate({ ids });
+          }}
+          disabled={isDisabled}
+        />
+      </CardContent>
+    </Card>
+  );
+};
+
+export default TransactionsClient;
