+++ app/layout.tsx
@@ -2,7 +2,7 @@ import { FullscreenLoader } from "@/components/fullscreen-loader";
 import { Toaster } from "@/components/ui/sonner";
 import { TooltipProvider } from "@/components/ui/tooltip";
 import { ClerkProvider, GoogleOneTap } from "@clerk/nextjs";
-import type { Metadata } from "next";
+import type { Metadata, Viewport } from "next";
 import { NextIntlClientProvider } from "next-intl";
 import { getMessages } from "next-intl/server";
 import { Inter } from "next/font/google";
@@ -12,7 +12,10 @@ import "./globals.css";
 const inter = Inter({ subsets: ["latin"] });
 
 export const metadata: Metadata = {
-  title: "Finance Pro",
+  title: {
+    template: "%s | Finance Pro",
+    default: "Finance Pro",
+  },
   description: "Finace Platform to track your money",
   icons: {
     icon: [
@@ -22,6 +25,13 @@ export const metadata: Metadata = {
       },
     ],
   },
+};
+
+export const viewport: Viewport = {
+  width: "device-width",
+  initialScale: 1,
+  maximumScale: 1,
+  userScalable: false,
 };
 
 export default async function RootLayout({
