+++ messages/en.json
@@ -19,6 +19,7 @@
     "Page": {
       "Header": "{key} Page"
     },
+    "DatePicker": "Pick A Date",
     "Action": {
       "Edit": "Edit",
       "Delete": "Delete",
